# Conflicts:
#	Explorer/Assets/DCL/AvatarRendering/Wearables/Systems/ResolveWearableByPointerSystem.cs
#	Explorer/Assets/DCL/AvatarRendering/Wearables/Wearables.asmdef
#	Explorer/Assets/DCL/Profiles/Self/Playground/SelfProfilePlayground.cs
#	Explorer/Assets/DCL/Profiles/Self/SelfProfile.cs
#	Explorer/Assets/Scripts/Global/Dynamic/DynamicWorldContainer.cs

+++ Explorer/Assets/DCL/PerformanceAndDiagnostics/Optimization/PerformanceBudgeting/Memory/ISystemMemory.cs
@@ -2,6 +2,6 @@ namespace DCL.Optimization.PerformanceBudgeting
 {
     public interface ISystemMemory
     {
-        ulong TotalSizeInMB { get; }
+        int TotalSizeInMB { get; }
     }
 }
